Created a services folder to store notion api accessing functions
